Bugfix Update 1.4.5
Hello there,

I am back again with another Bugfix Update. But first of all I have to apologize that it took so long to fix this bug, but sometimes, real life is way more important than something you are doing online.

Now to the bugfix, as many of you know, the plugin couldn't save data when the server crashed. That was because of the limitations of spigot on server crash. I added a way to store player data before the server is crashing.

So you now can enable backups! You can configure how often the data should be backed up and also select which data should be backed up. For performance and traffic reasons, the plugin will check if data has changed since the last backup, if not it will not perform a backup of that specific data.

  backup: # Backups of data are needed if the server may crash. If you disable it, data will be lost after a crash of the server.
    enabled: true # Set false to disable Backups of data and true to enable Backups of data.
    backupCycle: 1200 # The Cycle in which the plugin should Backup data, the time must be in ticks (1200 Ticks are 1 Minute)
    values:
      inventory: true # If the Inventory of the Player should be backed up.
      enderchest: true # If the Enderchest of the Player should be backed up.
      exp: true # If the EXP of the Player should be backed up.
      gamemode: true # If the Gamemode of the Player should be backed up.
      hunger: true # If the Hunger of the Player should be backed up.
      health: true # If the Health of the Player should be backed up.
      effects: false # If the Effects of the Player should be backed up.
      advancements: false # If the Advancements of the Player should be backed up.
      statistics: false # If the Statistics of the Player should be backed up.

If you update from any 1.4.x version you will have to add the following lines under the "settings:" section in the config. Or you can just backup the old config file and delete the old one.

Thanks again for everyone that reported this bug. And again, sorry for the long wait.

~ Leo