Bugfix Update 1.4.7
Good Evening,

I have another bugfix for y'all!
There was a bug, which rolled back the inventory of the player to the latest backup on server restart. It should now be fixed.

The next version will probably be 1.5.0!
However, the new features which will be implemented are way more complex and will take much ore time than any update before. If I would guess, it would take around 2 to 3 months. But there will be still support and bugfixes if necessary on the 1.4 gen of the plugin.

~ Leo