Update InventoryManager.java to include Java best practices
Hi, 

I looked through your code when searching for a sync plugin for my MC server and in your GitHub profile that you are currently learning Java. So, I thought I create this PR to show some best practices that are common in the Java and programming world. I hope that is okay for you.

First, I replaced the "[magic number](https://refactoring.guru/replace-magic-number-with-symbolic-constant)" `41` determining the size of the ItemStack array with a constant so the code becomes more readable. 

Then, instead of creating a new variable, using a while loop, and incrementing it manually at the end, the code can be made more concise when using a [for-loop](https://www.w3schools.com/java/java_for_loop.asp)

Finally, I added some blank lines between assignments, loops and statements in that method to make the code easier to read.