Folia and Bugfix Update 1.4.11
Hello there,

here is another Bugfix Update, but not only that. MySQL Sync has finally support for Folia Servers in the newest Minecraft Versions (1.20.x). Also we have a new Language again, this time Japanese from RedCatKnights and kosan_1014 big shoutout to them for translating the plugin.

Bugfixes:
When a player executed /sync reload and didn't have the correct permission, there was an error printed into the console.
When a player executed /sync and didn't have the correct permission, there was an error printed into the console.
(Both bugs were reported by こうさん thank you for reporting them to us.)

Also we are working on bringing our maven repository for the Developer API back up and running. It is online as a unsafe website as of yet, and with an older version of the MySQL Sync API.

~ Leo