Potion and Gamemode not Syncing Bug
This is a try to fix the bug, that Potioneffects and Gamemode were not syncing because the Plugin run these two syncs in async mode, which cannot be handled by neither Spigot/Paper or Folia. This should now be fixed. But DO NOT use this version on your productive Server, as it is not tested as of yet.